Working Score and Highscore System
Finished the score and highscore system to display current score and save the highscore on the computer. It accesses the saved value every game as soon as player spawns. Score is 1 point every second survived. Created a bool called scoreIncreasing to determine whether the score is increasing or not and it stops when the player dies and starts again when the player spawns.